+++ lib/src/integrations/test_coverage/command/dart/command/model/dart_test_coverage_config_data.dart
@@ -3,9 +3,11 @@ import 'package:impaktfull_cli/src/core/command/config/command_config.dart';
 class DartTestCoverageConfigData extends ConfigData {
   final bool runTests;
   final bool convertToLcov;
+  final List<RegExp> ignorePatterns;
 
   const DartTestCoverageConfigData({
-    this.runTests = true,
-    this.convertToLcov = true,
+    required this.runTests,
+    required this.convertToLcov,
+    required this.ignorePatterns,
   });
 }