+++ lib/src/integrations/test_coverage/command/flutter/command/flutter_test_coverage_command_config.dart
@@ -1,15 +1,30 @@
 import 'package:args/args.dart';
 import 'package:impaktfull_cli/src/core/command/config/command_config.dart';
+import 'package:impaktfull_cli/src/core/util/extensions/arg_result_extensions.dart';
 import 'package:impaktfull_cli/src/integrations/test_coverage/command/flutter/command/model/flutter_test_coverage_config_data.dart';
 
 class FlutterTestCoverageCommandConfig
     extends CommandConfig<FlutterTestCoverageConfigData> {
+  static const String _optionRunTests = 'runTests';
   const FlutterTestCoverageCommandConfig();
 
   @override
-  void addConfig(ArgParser argParser) {}
+  void addConfig(ArgParser argParser) {
+    argParser.addFlag(
+      _optionRunTests,
+      help: 'Run tests before generating test coverage report',
+      defaultsTo: true,
+    );
+  }
 
   @override
   FlutterTestCoverageConfigData parseResult(ArgResults? argResults) =>
-      FlutterTestCoverageConfigData();
+      FlutterTestCoverageConfigData(
+        runTests: argResults.getFlag(_optionRunTests),
+        ignorePatterns: [
+          RegExp(r'.*\.g\.dart$'),
+          RegExp(r'.*\.navigator\.dart$'),
+          RegExp(r'.*/injectable\.config\.dart$'),
+        ],
+      );
 }