+++ lib/src/integrations/test_coverage/command/flutter/command/model/flutter_test_coverage_config_data.dart
@@ -2,8 +2,10 @@ import 'package:impaktfull_cli/src/core/command/config/command_config.dart';
 
 class FlutterTestCoverageConfigData extends ConfigData {
   final bool runTests;
+  final List<RegExp> ignorePatterns;
 
   const FlutterTestCoverageConfigData({
-    this.runTests = true,
+    required this.runTests,
+    required this.ignorePatterns,
   });
 }